requirements.txt
# Dlogic Agent - Dependencies
anthropic>=0.40.0
python-telegram-bot>=21.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
line-bot-sdk>=3.0.0
flask>=3.0.0
gunicorn>=21.0.0
